Move UI classes to UI folder
