+++ app/src/main/java/com/example/instagramuicompose/ui/home/HomeViewModel.kt
@@ -1,11 +1,11 @@
-package com.example.instagramuicompose.reels
+package com.example.instagramuicompose.ui.home
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class ReelsViewModel @Inject constructor(
+class HomeViewModel @Inject constructor(
 
 ) : ViewModel() {
 }
