+++ app/src/main/java/com/example/instagramuicompose/ui/shop/ShopViewModel.kt
@@ -1,11 +1,11 @@
-package com.example.instagramuicompose.profile
+package com.example.instagramuicompose.ui.shop
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class ProfileViewModel @Inject constructor(
+class ShopViewModel @Inject constructor(
 
-) : ViewModel() {
+) : ViewModel()  {
 }
